+++ src/main/java/mx/com/nmp/pagos/mimonte/dao/conciliacion/ActividadRepository.java
@@ -38,7 +38,7 @@ public interface ActividadRepository extends JpaRepository<Actividad, Long> {
 	 * @return
 	 */
 	@Query("SELECT new mx.com.nmp.pagos.mimonte.dto.conciliacion.ConsultaActividadDTO(act.folio, act.fecha, act.descripcion) FROM Actividad act WHERE act.folio = :folio AND act.fecha BETWEEN :fechaDesde AND :fechaHasta  ORDER BY act.fecha DESC")
-	public List<ConsultaActividadDTO> findByFolioFechaDesdeAndFechaHasta(@Param("folio") final Integer folio,
+	public List<ConsultaActividadDTO> findByFolioFechaDesdeAndFechaHasta(@Param("folio") final Long folio,
 			@Param("fechaDesde") final Date fechaDesde, @Param("fechaHasta") final Date fechaHasta);
 
 	/**
@@ -50,7 +50,7 @@ public interface ActividadRepository extends JpaRepository<Actividad, Long> {
 	 * @return
 	 */
 	@Query("SELECT new mx.com.nmp.pagos.mimonte.dto.conciliacion.ConsultaActividadDTO(act.folio, act.fecha, act.descripcion) FROM Actividad act WHERE act.folio = :folio AND act.fecha >= :fechaDesde  ORDER BY act.fecha DESC")
-	public List<ConsultaActividadDTO> findByFolioAndFechaDesde(@Param("folio") final Integer folio,
+	public List<ConsultaActividadDTO> findByFolioAndFechaDesde(@Param("folio") final Long folio,
 			@Param("fechaDesde") final Date fechaDesde);
 
 	/**
@@ -62,7 +62,7 @@ public interface ActividadRepository extends JpaRepository<Actividad, Long> {
 	 * @return
 	 */
 	@Query("SELECT new mx.com.nmp.pagos.mimonte.dto.conciliacion.ConsultaActividadDTO(act.folio, act.fecha, act.descripcion) FROM Actividad act WHERE act.folio = :folio AND act.fecha <= :fechaHasta  ORDER BY act.fecha DESC")
-	public List<ConsultaActividadDTO> findByFolioAndFechaHasta(@Param("folio") final Integer folio,
+	public List<ConsultaActividadDTO> findByFolioAndFechaHasta(@Param("folio") final Long folio,
 			@Param("fechaHasta") final Date fechaHasta);
 
 	/**
@@ -73,6 +73,6 @@ public interface ActividadRepository extends JpaRepository<Actividad, Long> {
 	 * @return
 	 */
 	@Query("SELECT new mx.com.nmp.pagos.mimonte.dto.conciliacion.ConsultaActividadDTO(act.folio, act.fecha, act.descripcion) FROM Actividad act WHERE act.folio = :folio ORDER BY act.fecha DESC")
-	public List<ConsultaActividadDTO> findByFolio(@Param("folio") final Integer folio);
+	public List<ConsultaActividadDTO> findByFolio(@Param("folio") final Long folio);
 
 }
